Django RF/requirements.txt
Django==3.1.6
django-rest-framework==0.1.0
djangorestframework==3.12.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
django_cors_headers==3.7.0
